code/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
numpy==2.1.1
PyJWT==2.8.0
reverse_geocode==1.6.5
scipy==1.14.1
sqlparse==0.5.1
tzdata==2024.1
